fix(core): improve BusyBox ash POSIX compliance and script robustness

+++ decompressed/gui_file/etc/modgui_scripts/03_various.sh
@@ -5,11 +5,11 @@
 check_upgrade_shit() {
   logecho "Fixing sysupgrade if needed..."
 	if [ -f /lib/upgrade/resetgui.sh ]; then
-		rm /lib/upgrade/resetgui.sh
-		rm /lib/upgrade/transfer_bank1.sh
-		rm /lib/upgrade/upgradegui.sh
-		rm /lib/upgrade/checkver.sh
-		rm /lib/upgrade/hardreset.sh
+		rm -f /lib/upgrade/resetgui.sh
+		rm -f /lib/upgrade/transfer_bank1.sh
+		rm -f /lib/upgrade/upgradegui.sh
+		rm -f /lib/upgrade/checkver.sh
+		rm -f /lib/upgrade/hardreset.sh
 	fi
 }
 
@@ -18,26 +18,28 @@ restore_original_mapper() {
 	orig_dir=/rom/usr/share/transformer/mappings
 	target=/usr/share/transformer/mappings
 
-	if [ "$(md5sum $orig_dir/device2/Device.map | awk '{print $1}')" != "$(md5sum $target/device2/Device.map | awk '{print $1}')" ]; then
-		mkdir /tmp/tmp_bff_file
-		cp $target/bbf/VoiceService* /tmp/tmp_bff_file/
-		rm -r $target/bbf/*
-		rm -r $target/device2/*
-		rm -r $target/clash/*
-		rm -r $target/igd/*
-		cp $orig_dir/bbf/* $target/bbf
-		cp $orig_dir/clash/* $target/clash
-		cp $orig_dir/device2/* $target/device2
-		cp $orig_dir/igd/* $target/igd
-		cp /tmp/tmp_bff_file/* $target/bbf
-		rm -r /tmp/tmp_bff_file
+	if [ -f "$orig_dir/device2/Device.map" ] && [ -f "$target/device2/Device.map" ] && [ "$(md5sum "$orig_dir/device2/Device.map" 2>/dev/null | awk '{print $1}')" != "$(md5sum "$target/device2/Device.map" 2>/dev/null | awk '{print $1}')" ]; then
+		mkdir -p /tmp/tmp_bff_file
+		cp $target/bbf/VoiceService* /tmp/tmp_bff_file/ 2>/dev/null
+		rm -rf $target/bbf/*
+		rm -rf $target/device2/*
+		rm -rf $target/clash/*
+		rm -rf $target/igd/*
+		cp $orig_dir/bbf/* $target/bbf 2>/dev/null
+		cp $orig_dir/clash/* $target/clash 2>/dev/null
+		cp $orig_dir/device2/* $target/device2 2>/dev/null
+		cp $orig_dir/igd/* $target/igd 2>/dev/null
+		cp /tmp/tmp_bff_file/* $target/bbf 2>/dev/null
+		rm -rf /tmp/tmp_bff_file
 		logecho "Restored device-specific transformer mapping files"
 	fi
-	if [ ! -f $target/bbf/VoiceService.VoiceProfile.Line.map ]; then
-		cp $orig_dir/bbf/VoiceService.VoiceProfile.Line.map $orig_dir/bbf/
-	elif [ "$(md5sum $orig_dir/bbf/VoiceService.VoiceProfile.Line.map | awk '{print $1}')" != "$(md5sum $target/bbf/VoiceService.VoiceProfile.Line.map | awk '{print $1}')" ]; then
-		rm "$target/bbf/VoiceService.VoiceProfile.Line.map"
-		cp "$orig_dir/bbf/VoiceService.VoiceProfile.Line.map" "$orig_dir/bbf/"
+	if [ -f "$orig_dir/bbf/VoiceService.VoiceProfile.Line.map" ]; then
+		if [ ! -f "$target/bbf/VoiceService.VoiceProfile.Line.map" ]; then
+			cp "$orig_dir/bbf/VoiceService.VoiceProfile.Line.map" "$target/bbf/"
+		elif [ -f "$target/bbf/VoiceService.VoiceProfile.Line.map" ] && [ "$(md5sum "$orig_dir/bbf/VoiceService.VoiceProfile.Line.map" 2>/dev/null | awk '{print $1}')" != "$(md5sum "$target/bbf/VoiceService.VoiceProfile.Line.map" 2>/dev/null | awk '{print $1}')" ]; then
+			rm -f "$target/bbf/VoiceService.VoiceProfile.Line.map"
+			cp "$orig_dir/bbf/VoiceService.VoiceProfile.Line.map" "$target/bbf/"
+		fi
 	fi #Solve some problems with cwmp AGTEF_1.0.3
 
 	#Remove ignored Root device naming from transformer
@@ -50,11 +52,11 @@ transformer_lib_check() {
 	orig_dir=/rom/usr
 	target=/usr
 
-	if [ "$(md5sum $orig_dir/lib/lua/transformer/commitapply.lua | awk '{print $1}')" != "$(md5sum $target/lib/lua/transformer/commitapply.lua | awk '{print $1}')" ]; then
-		rm $target/share/transformer/mappings/rpc/*
-		rm $target/share/transformer/mappings/uci/*
-		rm -r $target/lib/lua/tch/*
-		rm -r $target/lib/lua/transformer/*
+	if [ -f "$orig_dir/lib/lua/transformer/commitapply.lua" ] && [ -f "$target/lib/lua/transformer/commitapply.lua" ] && [ "$(md5sum "$orig_dir/lib/lua/transformer/commitapply.lua" 2>/dev/null | awk '{print $1}')" != "$(md5sum "$target/lib/lua/transformer/commitapply.lua" 2>/dev/null | awk '{print $1}')" ]; then
+		rm -f $target/share/transformer/mappings/rpc/*
+		rm -f $target/share/transformer/mappings/uci/*
+		rm -rf $target/lib/lua/tch/*
+		rm -rf $target/lib/lua/transformer/*
 		cp $orig_dir/share/transformer/mappings/rpc/* $target/share/transformer/mappings/rpc
 		cp $orig_dir/share/transformer/mappings/uci/* $target/share/transformer/mappings/uci
 		cp $orig_dir/bin/transformer $target/bin/
@@ -71,8 +73,8 @@ transformer_lib_check() {
 			gui_pos=/tmp/GUI_dev.tar.bz2
 		fi
 		logecho "Found gui here: "$gui_pos
-		if [ $gui_pos != "" ] && [ -s $gui_pos ]; then
-			bzcat $gui_pos | tar -C / -xf - usr #reapply the upgrade as in the gui we store some of this file that we restored
+		if [ -n "$gui_pos" ] && [ -s "$gui_pos" ]; then
+			bzcat "$gui_pos" | tar -C / -xf - usr #reapply the upgrade as in the gui we store some of this file that we restored
 			logecho "Restoring transformer lib" #What is going on here? Doesn't even restart transformer???
 		fi
 		#/etc/init.d/transformer restart
@@ -81,11 +83,12 @@ transformer_lib_check() {
 
 check_wansensing() {
 	#Make sure that wansensing is under the correct dir
-	if [ -d /usr/lib/lua/wansensing ] && [ ! -d /usr/lib/lua/wansensingfw ] ; then
-    logecho "Old build detected, moving wansensing file"
-		rm /usr/lib/lua/wansensing/scripthelpers.lua
-		mv /usr/lib/lua/wansensingfw/scripthelpers.lua /usr/lib/lua/wansensing/scripthelpers.lua
-		rm -r /usr/lib/lua/wansensing
+	if [ -d /usr/lib/lua/wansensingfw ]; then
+		logecho "Old build detected, moving wansensing file"
+		[ -f /usr/lib/lua/wansensing/scripthelpers.lua ] && rm /usr/lib/lua/wansensing/scripthelpers.lua
+		[ ! -d /usr/lib/lua/wansensing ] && mkdir -p /usr/lib/lua/wansensing
+		[ -f /usr/lib/lua/wansensingfw/scripthelpers.lua ] && mv /usr/lib/lua/wansensingfw/scripthelpers.lua /usr/lib/lua/wansensing/scripthelpers.lua
+		rm -rf /usr/lib/lua/wansensingfw
 	fi
 }
 
@@ -94,9 +97,9 @@ create_simbolic_utility() {
 	if [ ! -h /usr/sbin/upgradegui ]; then
 		upgrade_utility=/usr/share/transformer/scripts/upgradegui
 		check_ver=/usr/share/transformer/scripts/checkver
-		if [ -f $upgrade_utility ]; then
-			ln -s $upgrade_utility /usr/sbin/upgradegui
-			ln -s $check_ver /usr/sbin/checkver
+		if [ -f "$upgrade_utility" ]; then
+			ln -sf "$upgrade_utility" /usr/sbin/upgradegui
+			ln -sf "$check_ver" /usr/sbin/checkver
 		fi
 	fi
 }
@@ -104,11 +107,11 @@ create_simbolic_utility() {
 update_checkver_upgrade_script() {
 	if [ -f /usr/share/transformer/scripts/upgradegui.sh ]; then
 	  logecho "Removing old update script"
-		rm /usr/sbin/upgradegui
-		rm /usr/sbin/checkver
+		rm -f /usr/sbin/upgradegui
+		rm -f /usr/sbin/checkver
 		create_simbolic_utility
-		rm /usr/share/transformer/scripts/upgradegui.sh
-		rm /usr/share/transformer/scripts/checkver.sh
+		rm -f /usr/share/transformer/scripts/upgradegui.sh
+		rm -f /usr/share/transformer/scripts/checkver.sh
 	fi
 }
 
@@ -137,7 +140,7 @@ cron_christmas() {
 }
 
 create_symlink() {
-	if [ ! -f /etc/rc.d/S70wol ]; then
+	if [ -f /etc/init.d/wol ] && [ ! -f /etc/rc.d/S70wol ]; then
 		/etc/init.d/wol enable
 	fi
 }
